+++ src/TestSuite/EventAssertTrait.php
@@ -36,6 +36,34 @@ trait EventAssertTrait
     }
 
     /**
+     * Asserts that an event was fired, with specified arguments
+     * @param string $eventName Event name
+     * @param array $eventArgs Event arguments
+     * @param \Tools\Event\EventDispatcher $EventDispatcher Event dispatcher
+     * @param string $message The message to use for failure
+     * @return void
+     * @since 1.5.10
+     */
+    public function assertEventFiredWithArgs(string $eventName, array $eventArgs, EventDispatcher $EventDispatcher, string $message = ''): void
+    {
+        $this->assertEventFired($eventName, $EventDispatcher, $message);
+
+        $matching = false;
+        foreach ($EventDispatcher->getEventList()->toArray() as $event) {
+            if ($event->getName() === $eventName && $event->getArgs() === $eventArgs) {
+                $matching = true;
+            }
+        }
+        if (!$message) {
+            $message = 'The `' . $eventName . '` event was not fired with the specified arguments';
+            if (count($eventArgs) === 1 && is_string(array_value_first($eventArgs))) {
+                $message .= ' `' . array_value_first($eventArgs) . '`';
+            }
+        }
+        $this->assertTrue($matching, $message);
+    }
+
+    /**
      * Asserts that an event was not fired
      * @param string $eventName Event name
      * @param \Tools\Event\EventDispatcher $EventDispatcher Event dispatcher